requirements.txt
flask
mysql-connector-python
waitress
python-dotenv